not run in view

+++ personalhelper/personalhelper/crawling/crawling/pipelines.py
@@ -3,7 +3,8 @@ from scraper.models import Comand
 
 
 class CrawlingPipeline(object):
-    Comand.objects.all().delete()
+    if Comand.objects.all():
+        Comand.objects.all().delete()
     
     def process_item(self, item, spider):
         adapter = ItemAdapter(item)
